create consts folder and move protocol-consts and reserved attribute definitions
