+++ src/formatters/__tests__/network-test.js
@@ -6,7 +6,7 @@ import {
 import { getEntityAttributes } from '../../utils/general';
 import {
   egoProperty, entityAttributesProperty, entityPrimaryKeyProperty, sessionProperty,
-} from '../../utils/reservedAttributes';
+} from '../../consts/reservedAttributes';
 
 // todo: add resequence ID test
 const protocolID = 123;
