+++ src/formatters/csv/__tests__/matrix-test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import { makeWriteableStream } from '../../../../config/setupTestEnv';
 import { AdjacencyMatrixFormatter, asAdjacencyMatrix } from '../matrix';
-import { ncSourceUUID, ncTargetUUID } from '../../../utils/reservedAttributes';
+import { ncSourceUUID, ncTargetUUID } from '../../../consts/reservedAttributes';
 
 const mockNetwork = (edges) => ({
   edges,
