+++ src/formatters/csv/edge-list.js
@@ -7,7 +7,7 @@ const {
   ncSourceUUID,
   ncTargetUUID,
   ncUUIDProperty,
-} = require('../../utils/reservedAttributes');
+} = require('../../consts/reservedAttributes');
 const { processEntityVariables } = require('../network');
 const { sanitizedCellValue, csvEOL } = require('./csv');
