+++ src/formatters/csv/ego-list.js
@@ -12,7 +12,7 @@ const {
   ncCaseProperty,
   ncSessionProperty,
   ncProtocolNameProperty,
-} = require('../../utils/reservedAttributes');
+} = require('../../consts/reservedAttributes');
 const { processEntityVariables } = require('../network');
 const { sanitizedCellValue, csvEOL } = require('./csv');
